requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.2.0

# Payment Processing
stripe==7.11.0

# Email
sendgrid==6.11.0
jinja2==3.1.3

# Analytics
user-agents==2.2.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
